2.6.0 release

+++ avt/Plotter/avtTransparencyActor.h
@@ -47,6 +47,7 @@
 
 #include <vector>
 #include <map>
+#include <cstring>
 
 class     vtkActor;
 class     vtkAppendPolyData;
@@ -194,7 +195,7 @@ class PLOTTER_API avtTransparencyActor
     bool                                             cachedTransparencies;
 
     void                                             SetUpActor(void);
-    void                                             PrepareDataset(int, int);
+    void                                             PrepareDataset(size_t, size_t);
     void                                             DetermineTransparencies();
 };
 #endif
